requirements.txt
requests==2.31.0
python-dotenv==1.0.0
sendgrid==6.10.0
feedparser==6.0.10
urllib3==2.0.7 